requirements.txt
Flask==3.1.0
langchain==0.3.14
langchain-chroma==0.1.4
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.41.1
waitress==3.0.2
Unidecode==1.3.8
pydantic==2.9.2
